requirements.txt
gensim
tqdm
nltk
sklearn
pandas
matplotlib
seaborn
simpletransformers
pytorch_metric_learning
hypopt
python-Levenshtein
